+++ src/style/index.ts
@@ -5,4 +5,5 @@ export {default as RFill} from './RFill';
 export {default as RCircle} from './RCircle';
 export {default as RRegularShape} from './RRegularShape';
 export {default as RText} from './RText';
+export {default as RBackground} from './RBackground';
 export {default as RIcon} from './RIcon';